DFE-1329 Group FilterItems in FullTable™ response for rendering FullTable™ and Permalinks
