--- src/GovUk.Education.ExploreEducationStatistics.Data.Api/ViewModels/Meta/TableBuilder/TableBuilderResultFilterItemMetaViewModel.cs
@@ -1,9 +1,0 @@
-namespace GovUk.Education.ExploreEducationStatistics.Data.Api.ViewModels.Meta.TableBuilder
-{
-    public class TableBuilderResultFilterItemMetaViewModel
-    {
-        public bool IsTotal { get; set; }
-        public string Label { get; set; }
-        public string Value { get; set; }
-    }
-}
